dp c-6

+++ store/models.py
@@ -6,14 +6,14 @@ from django.contrib.auth.models import User # Вроде как механика
 class Customer(models.Model):
     user = models.OneToOneField(User, null=True, blank=True, on_delete=models.CASCADE)
     name = models.CharField(max_length=200, null=True)
-    email = models.CharField(max_length=200, null=True)
+    email = models.CharField(max_length=200, null=True) # We can also use .EmailField(), mb this method better match to out pr
 
     def __str__(self):
         return self.name
 
 class Product(models.Model):
     name = models.CharField(max_length=200, null=True)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=7, decimal_places=2)
     digital = models.BooleanField(default=False, null=True, blank=True)
     image = models.ImageField(null=True, blank=True)
